+++ djson.go
@@ -8,6 +8,14 @@ import (
 	gov "github.com/asaskevich/govalidator"
 )
 
+type numbers interface {
+	uint | uint8 | uint16 | uint32 | uint64 | int | int8 | int16 | int32 | int64 | float32 | float64
+}
+
+type intergers interface {
+	uint | uint8 | uint16 | uint32 | uint64 | int | int8 | int16 | int32 | int64
+}
+
 const (
 	NULL   = 0
 	OBJECT = 1
